+++ Node.h
@@ -9,8 +9,8 @@ class Node
 public:
 	vector<Connection*> inputs;
 	vector<Connection*> outputs;
-	vector<float> weighted;
-	float normalized;
+	vector<double> weighted;
+	double normalized;
 
 	Node() : Node(vector<Connection*>(), vector<Connection*>()) {};
 	Node(vector<Connection*> inputs, vector<Connection*> outputs);
